Separated code for uber object and the rest of the functionality
